# cmd/apps/apis/rest/docs/swagger.yaml
basePath: /api/v1
definitions:
  cmd_apps_apis_rest_v1.ChangeParentRequest:
    properties:
      entry_uri:
        type: string
      exchange:
        type: boolean
      new_entry_uri:
        type: string
      replace:
        type: boolean
    required:
    - entry_uri
    - new_entry_uri
    type: object
  cmd_apps_apis_rest_v1.CreateEntryRequest:
    properties:
      filter:
        $ref: '#/definitions/v1.FilterConfig'
      kind:
        type: string
      rss:
        $ref: '#/definitions/v1.RssConfig'
      uri:
        type: string
    required:
    - uri
    type: object
  cmd_apps_apis_rest_v1.DeleteEntriesRequest:
    properties:
      uri_list:
        items:
          type: string
        type: array
    required:
    - uri_list
    type: object
  cmd_apps_apis_rest_v1.DeleteEntriesResponse:
    properties:
      deleted:
        items:
          type: string
        type: array
      message:
        type: string
    type: object
  cmd_apps_apis_rest_v1.DocumentProperty:
    properties:
      abstract:
        type: string
      author:
        type: string
      header_image:
        type: string
      keywords:
        items:
          type: string
        type: array
      marked:
        type: boolean
      notes:
        type: string
      publish_at:
        type: string
      source:
        type: string
      title:
        type: string
      unread:
        type: boolean
      url:
        type: string
      year:
        type: string
    type: object
  cmd_apps_apis_rest_v1.EntryDetail:
    properties:
      access:
        $ref: '#/definitions/v1.Access'
      access_at:
        type: string
      aliases:
        type: string
      changed_at:
        type: string
      created_at:
        type: string
      document:
        $ref: '#/definitions/cmd_apps_apis_rest_v1.DocumentProperty'
      entry:
        type: integer
      is_group:
        type: boolean
      kind:
        type: string
      modified_at:
        type: string
      name:
        type: string
      namespace:
        type: string
      property:
        $ref: '#/definitions/cmd_apps_apis_rest_v1.Property'
      size:
        type: integer
      storage:
        type: string
      uri:
        type: string
      version:
        type: integer
    type: object
  cmd_apps_apis_rest_v1.EntryInfo:
    properties:
      access_at:
        type: string
      changed_at:
        type: string
      created_at:
        type: string
      document:
        $ref: '#/definitions/cmd_apps_apis_rest_v1.DocumentProperty'
      entry:
        type: integer
      is_group:
        type: boolean
      kind:
        type: string
      modified_at:
        type: string
      name:
        type: string
      size:
        type: integer
      uri:
        type: string
    type: object
  cmd_apps_apis_rest_v1.FilterEntryRequest:
    properties:
      cel_pattern:
        type: string
      page:
        type: integer
      page_size:
        type: integer
    required:
    - cel_pattern
    type: object
  cmd_apps_apis_rest_v1.ListEntriesResponse:
    properties:
      entries:
        items:
          $ref: '#/definitions/cmd_apps_apis_rest_v1.EntryInfo'
        type: array
      pagination:
        $ref: '#/definitions/v1.PaginationInfo'
    type: object
...
  cmd_apps_apis_rest_v1.Property:
    properties:
      properties:
        additionalProperties:
          type: string
        type: object
      tags:
        items:
          type: string
        type: array
    type: object
...
  cmd_apps_apis_rest_v1.UpdateDocumentPropertyRequest:
    properties:
      marked:
        type: boolean
      unread:
        type: boolean
    type: object
  cmd_apps_apis_rest_v1.UpdateEntryRequest:
    properties:
      aliases:
        type: string
    type: object
  cmd_apps_apis_rest_v1.UpdatePropertyRequest:
    properties:
      properties:
        additionalProperties:
          type: string
        type: object
      tags:
        items:
          type: string
        type: array
    type: object
...
  v1.Access:
    properties:
      gid:
        type: integer
      permissions:
        items:
          type: string
        type: array
      uid:
        type: integer
    type: object
...
  v1.DocumentPropertyResponse:
    properties:
      properties:
        $ref: '#/definitions/cmd_apps_apis_rest_v1.DocumentProperty'
    type: object
  v1.EntryDetailResponse:
    properties:
      entry:
        $ref: '#/definitions/cmd_apps_apis_rest_v1.EntryDetail'
    type: object
  v1.EntryResponse:
    properties:
      entry:
        $ref: '#/definitions/cmd_apps_apis_rest_v1.EntryInfo'
    type: object
  v1.FilterConfig:
    properties:
      cel_pattern:
        type: string
    type: object
...
  v1.PaginationInfo:
    properties:
      page:
        type: integer
      page_size:
        type: integer
    type: object
  v1.PropertyResponse:
    properties:
      properties:
        $ref: '#/definitions/cmd_apps_apis_rest_v1.Property'
    type: object
  v1.RssConfig:
    properties:
      feed:
        type: string
      file_type:
        type: string
      site_name:
        type: string
      site_url:
        type: string
    type: object
...
host: localhost:7081
info:
  contact: {}
  description: NanaFS is a Reference Filing System API
  title: NanaFS API
  version: "1.0"
paths:
...
  /api/v1/entries:
    delete:
      consumes:
      - application/json
      description: Delete an entry by URI or ID
      parameters:
      - description: Entry URI
        in: query
        name: uri
        type: string
      - description: Entry ID
        in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.EntryResponse'
      summary: Delete entry
      tags:
      - Entries
    get:
      consumes:
      - application/json
      description: List all children entries of a directory
      parameters:
      - description: Entry URI
        in: query
        name: uri
        type: string
      - description: Entry ID
        in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd_apps_apis_rest_v1.ListEntriesResponse'
      summary: List entry children
      tags:
      - Entries
    post:
      consumes:
      - application/json
      description: Create a new entry with specified URI, kind and optional configuration
      parameters:
      - description: Entry creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/cmd_apps_apis_rest_v1.CreateEntryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/v1.EntryResponse'
      summary: Create a new entry
      tags:
      - Entries
    put:
      consumes:
      - application/json
      description: Update an existing entry's aliases
      parameters:
      - description: Update entry request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/cmd_apps_apis_rest_v1.UpdateEntryRequest'
      - description: Entry URI
        in: query
        name: uri
        type: string
      - description: Entry ID
        in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.EntryDetailResponse'
      summary: Update entry
      tags:
      - Entries
  /api/v1/entries/batch-delete:
    post:
      consumes:
      - application/json
      description: Batch delete multiple entries by URI list
      parameters:
      - description: Delete entries request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/cmd_apps_apis_rest_v1.DeleteEntriesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd_apps_apis_rest_v1.DeleteEntriesResponse'
      summary: Delete entries
      tags:
      - Entries
  /api/v1/entries/details:
    get:
      consumes:
      - application/json
      description: Get detailed information of an entry by URI or ID
      parameters:
      - description: Entry URI
        in: query
        name: uri
        type: string
      - description: Entry ID
        in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.EntryDetailResponse'
      summary: Get entry details
      tags:
      - Entries
  /api/v1/entries/document:
    get:
      consumes:
      - application/json
      description: Get document properties of an entry
      parameters:
      - description: Entry URI
        in: query
        name: uri
        type: string
      - description: Entry ID
        in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.DocumentPropertyResponse'
      summary: Get entry document
      tags:
      - Entries
    put:
      consumes:
      - application/json
      description: Update document properties (unread, marked) of an entry
      parameters:
      - description: Update document request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/cmd_apps_apis_rest_v1.UpdateDocumentPropertyRequest'
      - description: Entry URI
        in: query
        name: uri
        type: string
      - description: Entry ID
        in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.DocumentPropertyResponse'
      summary: Update document property
      tags:
      - Entries
  /api/v1/entries/parent:
    put:
      consumes:
      - application/json
      description: Move an entry to a new parent directory
      parameters:
      - description: Change parent request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/cmd_apps_apis_rest_v1.ChangeParentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.EntryResponse'
      summary: Change entry parent
      tags:
      - Entries
  /api/v1/entries/property:
    get:
      consumes:
      - application/json
      description: Get properties and tags of an entry
      parameters:
      - description: Entry URI
        in: query
        name: uri
        type: string
      - description: Entry ID
        in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.PropertyResponse'
      summary: Get entry property
      tags:
      - Entries
    put:
      consumes:
      - application/json
      description: Update properties and tags of an entry
      parameters:
      - description: Update property request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/cmd_apps_apis_rest_v1.UpdatePropertyRequest'
      - description: Entry URI
        in: query
        name: uri
        type: string
      - description: Entry ID
        in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.PropertyResponse'
      summary: Update entry property
      tags:
      - Entries
  /api/v1/entries/search:
    post:
      consumes:
      - application/json
      description: Filter entries using CEL pattern matching
      parameters:
      - description: Filter entry request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/cmd_apps_apis_rest_v1.FilterEntryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd_apps_apis_rest_v1.ListEntriesResponse'
      summary: Filter entries
      tags:
      - Entries
...
swagger: "2.0"